Linux fix 1/?

+++ Classes/ResourceManager/ResourceManager.cpp
@@ -259,7 +259,7 @@ FontData ResourceManager::loadFontFromFile(const GLchar *file)
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			(GLuint)face->glyph->advance.x
 		};
 
 		characterMap.insert(std::pair <GLchar, FontData::Character> (c, character));
